Block devices & MM collaboration (#28)
* Add parameter checks for Virt

* Add built-in virt context switching & TLB flushing

* Fix TLB

* Initialize devices (unfinished)

* Add VirtIO HAL implementation

* Add common block device trait

* Initialize devices

* Optimize device init strategy

* Add basic user tasks

* Add phys backend for block devices

* Add block backend conversions

* Fix shared mappings (unfinished)

* Add pinning for address mapping

+++ mizu/lib/kmem/src/frame.rs
@@ -8,7 +8,7 @@ use core::{
 use rv39_paging::{LAddr, PageAlloc, PAGE_MASK, PAGE_SHIFT, PAGE_SIZE};
 use static_assertions::const_assert_eq;
 
-const COUNT_SHIFT: usize = 5;
+const COUNT_SHIFT: u32 = 5;
 const MAX_COUNT: usize = 1 << (PAGE_SHIFT - COUNT_SHIFT);
 const COUNT_MASK: usize = (MAX_COUNT - 1) << COUNT_SHIFT;
